Only create symlinks if the Linux LaTeX dir exists
Avoids issues on non-Linux OSes

+++ linux/custom.profile
@@ -5,7 +5,7 @@ export JAVA_HOME
 export GRADLE_HOME
 export PATH="$JAVA_HOME/bin:$GRADLE_HOME/bin:$LATEX_HOME:$PATH"
 
-if ! [[ -f /usr/local/bin/pdflatex ]]; then
+if ! [[ -f /usr/local/bin/pdflatex ]] && realpath $LATEX_HOME &>/dev/null; then
     sudo ln -s $LATEX_HOME/* /usr/local/bin/
 fi
